pyproject.toml
[project]
name = "market-analyst"
version = "0.1.0"
description = "Multi-agent Indian stock market analyst using LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4.0",
    "langchain-groq>=1.0.0",
    "yfinance>=0.2.40",
    "ddgs>=9.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "streamlit>=1.40.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
